Time: 241 ms (44.88%) | Memory: 73.8 MB (48.80%) - LeetSync

+++ 2882-ways-to-express-an-integer-as-sum-of-powers/ways-to-express-an-integer-as-sum-of-powers.cpp
@@ -62,13 +62,13 @@ ll f(int i,ll rem,int &x,ll &largestBase,vector<vector<ll>> &dp)
             {
                 int cur=pow(i,x);
 
-                ll take=0;
+                int take=0;
                 if(cur <= rem)
                 take=dp[i+1][rem-cur];
 
-                ll notTake=dp[i+1][rem];
+                int notTake=dp[i+1][rem];
 
-                dp[i][rem]=(take+notTake)%mod;
+                dp[i][rem]=((ll)take+notTake)%mod;
             }
         }
